Testing new parameterisation
This parameterisation returns the graph that is the closest to the
original data that I have found so far, however it also has a runtime
error, no doubt due to the large value of parameter a. I don't know how
to prevent the error but I thought I should keep a copy of this
parameterisation on a separate script.